Move eslint specific code over to razzle-plugin-eslint
