+++ packages/razzle-plugin-eslint/index.js
@@ -1,8 +1,28 @@
 'use strict';
 
+const eslintFormatter = require('react-dev-utils/eslintFormatter');
+
 module.exports = function modify(config) {
-  console.log(
-    'Not yet implemented. Help out at https://github.com/jaredpalmer/razzle'
-  );
+  const mainEslintOptions = {
+    baseConfig: {
+      extends: [require.resolve('eslint-config-react-app')],
+    },
+    formatter: eslintFormatter,
+    eslintPath: require.resolve('eslint'),
+    ignore: false,
+    useEslintrc: true,
+  };
+
+  config.module.rules = [
+    {
+      test: /\.(js|jsx|mjs)$/,
+      enforce: 'pre',
+      loader: require.resolve('eslint-loader'),
+      options: mainEslintOptions,
+      exclude: /node_modules/,
+    },
+    ...config.module.rules,
+  ];
+
   return config;
 };
