+++ packages/razzle-plugin-eslint/package.json
@@ -2,5 +2,15 @@
   "name": "razzle-plugin-eslint",
   "version": "2.2.0",
   "main": "index.js",
-  "license": "MIT"
+  "license": "MIT",
+  "dependencies": {
+    "babel-eslint": "^10.0.1",
+    "eslint": "5.6.1",
+    "eslint-config-react-app": "^3.0.3",
+    "eslint-loader": "^2.1.1",
+    "eslint-plugin-flowtype": "^2.50.3",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1"
+  }
 }
