+++ packages/razzle/config/paths.js
@@ -54,7 +54,6 @@ module.exports = {
   appClientIndexJs: resolveApp('src/client'),
   testsSetup: resolveApp('src/setupTests.js'),
   appBabelRc: resolveApp('.babelrc'),
-  appEslintRc: resolveApp('.eslintrc'),
   appRazzleConfig: resolveApp('razzle.config.js'),
   nodePaths: nodePaths,
   ownPath: resolveOwn('.'),
