+++ packages/razzle/package.json
@@ -31,20 +31,12 @@
     "@babel/core": "^7.1.2",
     "assets-webpack-plugin": "^3.9.7",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
     "babel-loader": "^8.0.4",
     "babel-preset-razzle": "^3.0.0-alpha.2",
     "chalk": "^2.4.1",
     "css-loader": "^1.0.0",
     "dotenv": "6.0.0",
-    "eslint": "5.6.1",
-    "eslint-config-react-app": "^3.0.3",
-    "eslint-loader": "^2.1.1",
-    "eslint-plugin-flowtype": "^2.50.3",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-react": "^7.11.1",
     "file-loader": "^3.0.1",
     "fs-extra": "7.0.0",
     "jest": "^23.6.0",
